ffff

+++ device/Cons/Constants.cs
@@ -7,9 +7,9 @@
         public const int MAX_LENGTH_HARD_DRIVER = 50;
         public const int MAX_WEIGHT = 30;
         public const int MAX_RAM = 512;
-        public const double MAX_PIN = 3000;
+        public const decimal MAX_PIN = 3000;
         public const int MIN_RAM = 2;
-        public const double MAX_PRICE = 100000000;
+        public const decimal MAX_PRICE = 100000000;
         public const int MAX_QUANTITY = 999;
         public const int MAX_REPEAT_NAME = 5;
     }
