+++ device/DTO/Laptop/CreateLaptop.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public int IdProducer { get; set; }
-        public double CostPrice { get; set; }
-        public double SoldPrice { get; set; }
+        public decimal CostPrice { get; set; }
+        public decimal SoldPrice { get; set; }
     }
 }
