+++ device/DTO/LaptopDetail/CreateLaptopDetail.cs
@@ -9,11 +9,11 @@
         public string HardDriver { get; set; }
         public int IdMonitor { get; set; }
         public string Webcam { get; set; }
-        public double Weight { get; set; }
-        public double Height { get; set; }
-        public double Width { get; set; }
-        public double Length { get; set; }
-        public double BatteryCapacity { get; set; }
+        public decimal Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Width { get; set; }
+        public decimal Length { get; set; }
+        public decimal BatteryCapacity { get; set; }
         public byte[]? Image { get; set; }
         public int idLaptop { get; set; }
     }
